refactor: relocate components to /components - Move components from app/components to /components - Remove test console.log from EventCarousel
