+++ client/components/NavBar.tsx
